Remove dependency on terraform sdk
Versions of the terraform-plugin-sdk >= v2.0.0 made the hashcode
implementation module private.

I copied here since this is a helper function that

0. doesn't need to stay in-sync with the implementation in
terraform-plugin-sdk; and
1. isn't likely to change anyway

This also allows us to drop the terraform-plugin-sdk dependency, which
felt like a leaky abstraction and might have prevented its use in
applications other than for terraform providers.